+++ make/inc/customizer/fonts.php
@@ -218,11 +218,13 @@ function ttfmake_get_google_font_subsets() {
 	return apply_filters( 'make_get_google_font_subsets', array(
 		'all'          => __( 'All', 'make' ),
 		'arabic'       => __( 'Arabic', 'make' ),
+		'bengali'      => __( 'Bengali', 'make' ),
 		'cyrillic'     => __( 'Cyrillic', 'make' ),
 		'cyrillic-ext' => __( 'Cyrillic Extended', 'make' ),
 		'devanagari'   => __( 'Devanagari', 'make' ),
 		'greek'        => __( 'Greek', 'make' ),
 		'greek-ext'    => __( 'Greek Extended', 'make' ),
+		'gujarati'     => __( 'Gujarati', 'make' ),
 		'hebrew'       => __( 'Hebrew', 'make' ),
 		'khmer'        => __( 'Khmer', 'make' ),
 		'latin'        => __( 'Latin', 'make' ),